requirements.txt
# requirements.txt

# --- Core framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database & migrations ---
SQLAlchemy==2.0.34
psycopg2-binary
alembic==1.13.3

# --- Auth / Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- Background tasks / broker ---
celery==5.4.0
redis==5.1.1

# --- HTTP client for security checks ---
requests==2.32.3

# --- PDF generation & templating ---
Jinja2==3.1.4
weasyprint==62.3

# --- Pydantic & settings ---
pydantic>=2.0.0
pydantic-settings
python-dotenv==1.0.1

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2
python-multipart